refactor(parser): BottleFile using new SeabirdDataFile parent class
