+++ src/ctdam/parser/seabird_data_files.py
@@ -1,8 +1,9 @@
+import io
 import logging
 import re
-from datetime import datetime, timedelta
+from datetime import datetime, time, timedelta
 from pathlib import Path
-from typing import Tuple
+from typing import Tuple, Union
 
 import numpy as np
 import odf.sbe.accessors
@@ -90,10 +91,6 @@ class SeabirdDataFile:
         self.only_header = only_header
         self.read_file()
         self.metadata = self.structure_metadata(self.custom_metadata)
-        # if len(self.sensor_metadata) > 0:
-        #     self.sensors = self.sensor_xml_to_flattened_dict(
-        #         "".join(self.sensor_metadata)
-        #     )
         self.start_time = self.reading_start_time()
         self.start_position = self.reading_start_position()
         self.read_event_information()
@@ -645,3 +642,149 @@ class BottleLogFile(SeabirdDataFile):
         df["start_range"].astype("int")
         df["end_range"].astype("int")
         return df
+
+
+class BottleFile(SeabirdDataFile):
+    """
+    Class that represents a Sea-Bird Bottle File (.btl) .
+
+    Organizes the files table information into a pandas dataframe. This
+    allows the usage of this powerful library for statistics, visualization,
+    data manipulation, export, etc.
+
+    Parameters
+    ----------
+    path_to_file : Path | str
+        The path to the .btl file
+    only_header : bool
+        Whether to only check the header and not parse data
+    """
+
+    def __init__(self, path_to_file: Path | str):
+        super().__init__(path_to_file)
+        self.df = self.create_dataframe()
+        self.adding_timestamp_column()
+
+    def create_dataframe(self):
+        """
+        Creates a dataframe out of the .btl file.
+        Handles the double data header correctly.
+        """
+        top_names, bottom_names = self.reading_data_header()
+        # creating statistics column to store the row type information:
+        # 4 rows per bottle, average, standard deviation, max value, min value
+        top_names.append("Statistic")
+        data_text = "".join(self.raw_data[2:])
+
+        df = pd.read_fwf(
+            io.StringIO(data_text),
+            index_col=False,
+            header=None,
+            names=top_names,
+        )
+
+        # handling the double row header
+        rowtypes = df[df.columns[-1]].unique()
+
+        def separate_double_header_row(df, column, length):
+            """
+            Differentiates the two header rows.
+            """
+            column_idx = df.columns.get_loc(column)
+            old_column = df.iloc[::length, column_idx].reset_index(drop=True)
+            new_column = df.iloc[1::length, column_idx].reset_index(drop=True)
+            old_column_expanded = pd.Series(
+                np.repeat(old_column, length)
+            ).reset_index(drop=True)
+            new_column_expanded = pd.Series(
+                np.repeat(new_column, length)
+            ).reset_index(drop=True)
+            df[column] = old_column_expanded
+            df.insert(
+                column_idx + 1, bottom_names[column_idx], new_column_expanded
+            )
+            return df
+
+        df = separate_double_header_row(df, "Date", len(rowtypes))
+        df = separate_double_header_row(df, top_names[0], len(rowtypes))
+        # remove brackets around statistics values
+        df["Statistic"] = df["Statistic"].str.strip("()")
+        df = df.rename(
+            mapper={"Btl_ID": "Bottle_ID", "Bottle": "Bottle_ID"}, axis=1
+        )
+        return df
+
+    def adding_timestamp_column(self):
+        """
+        Creates a timestamp column that holds both, Date and Time information.
+        """
+        # constructing timestamp column
+        self.df.Date = pd.to_datetime(self.df.Date)
+        timestamp = []
+        for datepoint, timepoint in zip(self.df.Date, self.df.Time):
+            timestamp.append(
+                datetime.combine(
+                    datepoint,
+                    time.fromisoformat(str(timepoint)),
+                ).timestamp()
+            )
+        self.df.insert(2, "unixtime", timestamp)
+
+    def selecting_rows(
+        self,
+        df=None,
+        statistic_of_interest: Union[list, str] = ["avg"],
+    ):
+        """
+        Creates a dataframe with the given row identifier, using the
+        statistics column. A single string or a list of strings can be
+        processed.
+
+        Parameters
+        ----------
+        df : pandas.Dataframe :
+            the files Pandas representation (Default value = self.df)
+        statistic_of_interest : list or str
+            collection of values of the 'statistics' column in self.df
+            (Default value = ['avg'])
+        statistic_of_interest: Union[list
+
+        str] :
+             (Default value = ["avg"])
+
+        Returns
+        -------
+
+
+        """
+        df = self.df if df is None else df
+        # ensure that the input is a list, so that isin() can do its job
+        if isinstance(statistic_of_interest, str):
+            statistic_of_interest = [statistic_of_interest]
+        self.df = df.loc[df["Statistic"].isin(statistic_of_interest)]
+        self.df.drop(columns=["Statistic"], inplace=True)
+
+    def reading_data_header(self):
+        """
+        Identifies and separatly collects the rows that specify the data
+        tables headers.
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+
+
+        """
+        n = 11  # fix column width of a seabird btl file
+        top_line = self.raw_data[0]
+        second_line = self.raw_data[1]
+        top_names = [
+            top_line[i : i + n].split()[0]
+            for i in range(0, len(top_line) - n, n)
+        ]
+        bottom_names = [
+            second_line[i : i + n].split()[0] for i in range(0, 2 * n, n)
+        ]
+        return top_names, bottom_names
